+++ engine/MapFunctions.ts
@@ -1,22 +1,22 @@
 import { getStartCoords } from './getStartCoords'
 import { MapComponentType } from './MapComponent'
-import { addComponent, getComponent, hasComponent } from '@xrengine/engine/src/ecs/functions/ComponentFunctions'
-import { DebugNavMeshComponent } from '@xrengine/engine/src/debug/DebugNavMeshComponent'
-import { Entity } from '@xrengine/engine/src/ecs/classes/Entity'
+import { addComponent, getComponent, hasComponent } from '@etherealengine/engine/src/ecs/functions/ComponentFunctions'
+import { DebugNavMeshComponent } from '@etherealengine/engine/src/debug/DebugNavMeshComponent'
+import { Entity } from '@etherealengine/engine/src/ecs/classes/Entity'
 import { Object3D, Group, Mesh } from 'three'
-import { Engine } from '@xrengine/engine/src/ecs/classes/Engine'
-import { NavMeshComponent } from '@xrengine/engine/src/navigation/component/NavMeshComponent'
+import { Engine } from '@etherealengine/engine/src/ecs/classes/Engine'
+import { NavMeshComponent } from '@etherealengine/engine/src/navigation/component/NavMeshComponent'
 import { MapAction, mapReducer } from './MapReceptor'
 import { MapComponent } from './MapComponent'
 import { getPhases, startPhases } from './functions/PhaseFunctions'
-import { LoadGLTF } from '@xrengine/engine/src/assets/functions/LoadGLTF'
-import { avatarHalfHeight } from '@xrengine/engine/src/avatar/functions/createAvatar'
+import { LoadGLTF } from '@etherealengine/engine/src/assets/functions/LoadGLTF'
+import { avatarHalfHeight } from '@etherealengine/engine/src/avatar/functions/createAvatar'
 import { Text } from 'troika-three-text'
-import { Object3DComponent } from '@xrengine/engine/src/scene/components/Object3DComponent'
-import { ComponentJson } from '@xrengine/common/src/interfaces/SceneInterface'
-import { isClient } from '@xrengine/engine/src/common/functions/isClient'
-import { registerSceneLoadPromise } from '@xrengine/engine/src/scene/functions/SceneLoading'
-import { EntityNodeComponent } from '@xrengine/engine/src/scene/components/EntityNodeComponent'
+import { Object3DComponent } from '@etherealengine/engine/src/scene/components/Object3DComponent'
+import { ComponentJson } from '@etherealengine/common/src/interfaces/SceneInterface'
+import { isClient } from '@etherealengine/engine/src/common/functions/isClient'
+import { registerSceneLoadPromise } from '@etherealengine/engine/src/scene/functions/SceneLoading'
+import { EntityNodeComponent } from '@etherealengine/engine/src/scene/components/EntityNodeComponent'
 import { addChildFast, setPosition } from './util'
 import { debounce } from 'lodash'
