+++ engine/MapUpdateSystem.ts
@@ -1,20 +1,20 @@
-import { System } from '@xrengine/engine/src/ecs/classes/System'
+import { System } from '@etherealengine/engine/src/ecs/classes/System'
 import { MapComponent } from './MapComponent'
-import { TransformComponent } from '@xrengine/engine/src/transform/components/TransformComponent'
+import { TransformComponent } from '@etherealengine/engine/src/transform/components/TransformComponent'
 import { fromMetersFromCenter, LongLat } from './functions/UnitConversionFunctions'
 import { addChildFast, multiplyArray, setPosition, vectorToArray } from './util'
 import { Vector3 } from 'three'
-import { Entity } from '@xrengine/engine/src/ecs/classes/Entity'
-import { Object3DComponent } from '@xrengine/engine/src/scene/components/Object3DComponent'
-import { getComponent, defineQuery, addComponent } from '@xrengine/engine/src/ecs/functions/ComponentFunctions'
-import { World } from '@xrengine/engine/src/ecs/classes/World'
+import { Entity } from '@etherealengine/engine/src/ecs/classes/Entity'
+import { Object3DComponent } from '@etherealengine/engine/src/scene/components/Object3DComponent'
+import { getComponent, defineQuery, addComponent } from '@etherealengine/engine/src/ecs/functions/ComponentFunctions'
+import { World } from '@etherealengine/engine/src/ecs/classes/World'
 import isIntersectCircleCircle from './functions/isIntersectCircleCircle'
-import { AvatarComponent } from '@xrengine/engine/src/avatar/components/AvatarComponent'
-import { NavMeshComponent } from '@xrengine/engine/src/navigation/component/NavMeshComponent'
+import { AvatarComponent } from '@etherealengine/engine/src/avatar/components/AvatarComponent'
+import { NavMeshComponent } from '@etherealengine/engine/src/navigation/component/NavMeshComponent'
 import { accessMapState } from './MapReceptor'
 import { Downgraded } from '@hookstate/core'
 import { getPhases, startPhases, resetPhases } from './functions/PhaseFunctions'
-import { TargetCameraRotationComponent } from '@xrengine/engine/src/camera/components/TargetCameraRotationComponent'
+import { TargetCameraRotationComponent } from '@etherealengine/engine/src/camera/components/TargetCameraRotationComponent'
 
 const PI2 = Math.PI * 2
 const $vector3 = new Vector3()
