+++ engine/functions/computeTileBoundingBox.ts
@@ -1,6 +1,6 @@
 import { BBox } from 'geojson'
 import { TILE_ZOOM } from '../constants'
-import { LongLat, tileXToLong, tileYToLat, toMetersFromCenter } from '../functions/UnitConversionFunctions'
+import { LongLat, tileXToLong, tileYToLat, toMetersFromCenter } from './UnitConversionFunctions'
 
 const $array2 = Array(2)
