+++ tests/MapUpdateSystem.test.ts
@@ -1,17 +1,17 @@
 import assert from 'assert'
 import sinon, { SinonSpy } from 'sinon'
 import mock from 'mock-require'
-import { createWorld, World } from '@xrengine/engine/src/ecs/classes/World'
+import { createWorld, World } from '@etherealengine/engine/src/ecs/classes/World'
 import { Group, Mesh, Object3D, Quaternion, Vector3 } from 'three'
-import { createEntity } from '@xrengine/engine/src/ecs/functions/EntityFunctions'
-import { addComponent, getComponent } from '@xrengine/engine/src/ecs/functions/ComponentFunctions'
-import { TransformComponent } from '@xrengine/engine/src/transform/components/TransformComponent'
-import { Entity } from '@xrengine/engine/src/ecs/classes/Entity'
-import { Object3DComponent } from '@xrengine/engine/src/scene/components/Object3DComponent'
+import { createEntity } from '@etherealengine/engine/src/ecs/functions/EntityFunctions'
+import { addComponent, getComponent } from '@etherealengine/engine/src/ecs/functions/ComponentFunctions'
+import { TransformComponent } from '@etherealengine/engine/src/transform/components/TransformComponent'
+import { Entity } from '@etherealengine/engine/src/ecs/classes/Entity'
+import { Object3DComponent } from '@etherealengine/engine/src/scene/components/Object3DComponent'
 import { lineString } from '@turf/helpers'
-import { System } from '@xrengine/engine/src/ecs/classes/System'
-import { AvatarComponent } from '@xrengine/engine/src/avatar/components/AvatarComponent'
-import { NavMeshComponent } from '@xrengine/engine/src/navigation/component/NavMeshComponent'
+import { System } from '@etherealengine/engine/src/ecs/classes/System'
+import { AvatarComponent } from '@etherealengine/engine/src/avatar/components/AvatarComponent'
+import { NavMeshComponent } from '@etherealengine/engine/src/navigation/component/NavMeshComponent'
 import { MapAction, mapReducer } from '../engine/MapReceptor'
 import { MapStateUnwrapped } from '../engine/types'
 import { MapComponent } from '../engine/MapComponent'
