+++ xrengine.config.ts
@@ -1,8 +1,8 @@
-import type { ProjectConfigInterface } from '@xrengine/projects/ProjectConfigInterface'
+import type { ProjectConfigInterface } from '@etherealengine/projects/ProjectConfigInterface'
 
 const config: ProjectConfigInterface = {
   onEvent: undefined,
-  thumbnail: '/static/xrengine_thumbnail.jpg',
+  thumbnail: '/static/etherealengine_thumbnail.jpg',
   worldInjection: () => import('./worldInjection'),
   routes: {},
   services: undefined,
